requirements.txt
streamlit>=1.36.0
pdfplumber>=0.11.0
python-docx>=1.1.0
pyyaml>=6.0.1